casandra_project/requirements.txt
arrow==0.15.5
asgiref==3.2.3
cassandra-driver==3.19.0
cassandra-migrate==0.1.2
Click==7.0
Django==2.2.9
django-cassandra-engine==1.5.5
future==0.18.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.13
six==1.13.0
sqlparse==0.3.0
tabulate==0.8.6
